Refactor Swiboe class
1. Move Swiboe class to swiboe.rs.
2. Use swiboe::spawn to generate swiboe spinning thread.
