+++ src/plugin_core.rs
@@ -4,7 +4,7 @@
 
 use ::ipc_bridge;
 use ::rpc;
-use ::server;
+use ::swiboe;
 use serde_json;
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -14,11 +14,11 @@ pub struct NewRpcRequest {
 }
 
 pub struct CorePlugin {
-    commands: server::CommandSender,
+    commands: swiboe::SenderTo,
 }
 
 impl CorePlugin {
-    pub fn new(commands: server::CommandSender) -> Self {
+    pub fn new(commands: swiboe::SenderTo) -> Self {
         CorePlugin {
             commands: commands,
         }
@@ -27,7 +27,7 @@ impl CorePlugin {
     pub fn call(&self, caller: ipc_bridge::ClientId, rpc_call: &rpc::Call) -> rpc::Result {
         match &rpc_call.function as &str {
             "core.exit" => {
-                self.commands.send(server::Command::Quit).unwrap();
+                self.commands.send(swiboe::Command::Quit).unwrap();
                 rpc::Result::success(())
             },
             // NOCOM(#sirver): These args can be pulled out into Serializable structs.
@@ -39,7 +39,7 @@ impl CorePlugin {
                 };
 
                 self.commands.send(
-                    server::Command::NewRpc(caller, args.name, args.priority)).unwrap();
+                    swiboe::Command::NewRpc(caller, args.name, args.priority)).unwrap();
                 rpc::Result::success(())
             },
             // NOCOM(#sirver): this should not panic, but return an error.
